move html rewriting functions to util.ts
